Revert "finish all product page (makeshift code)"
