--- src/assets/productData/productDetailData.js
@@ -1,16 +1,0 @@
-import categoryData from "./categoryData";
-
-const productDetailData = () => {
-  const data = {};
-  // "[product-link]": "[data]"
-
-  for (const key in categoryData) {
-    categoryData[key].productList.forEach((item) => {
-      data[item.productPath] = item;
-    });
-  }
-
-  return data;
-};
-
-export default productDetailData;
