--- src/assets/productData/text.txt
@@ -1,1 +1,0 @@
-Thể tích: 750ml/29% ALC.;Quy cách đóng gói: Như trên bao bì.
