+++ src/layouts/components/Header/Header.module.scss
@@ -225,7 +225,6 @@
           left: 0;
           transform-origin: calc(40px + 15px) 0;
           animation: grow-up-animation 0.3s ease-out forwards;
-          z-index: 9999999;
 
           .toggle-link-container {
             width: 260px;
