+++ src/pages/ProductCategory/ProductCategory.js
@@ -1,6 +1,6 @@
 import { useParams } from "react-router-dom";
 import { useMemo } from "react";
-import categoryData from "assets/productData/categoryData";
+import productData from "assets/productData/productData";
 import classNames from "classnames/bind";
 import style from "./ProductCategory.module.scss";
 import ProductContainer from "components/ProductContainer/ProductContainer";
@@ -10,16 +10,16 @@ const cx = classNames.bind(style);
 function ProductCategory() {
   const { categoryId } = useParams();
 
-  const category = useMemo(() => {
-    return categoryData[categoryId];
+  const categoryData = useMemo(() => {
+    return productData[categoryId];
   }, [categoryId]);
 
   return (
     <div className={cx("container")}>
       <div className={cx("product-content-wrapper")}>
         <ProductContainer 
-          headerTitle={category.title}
-          product={category.productList}
+          headerTitle={categoryData.title}
+          product={categoryData.information}
         />
       </div>
     </div>
