--- src/pages/ProductDetail/ProductDetail.module.scss
@@ -1,10 +1,0 @@
-:root {
-  --content-wrapper-max-width: 1170px;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding-top: 60px;
-}
