--- src/pages/ProductDetail/components/ProductOverview/ProductOverview.module.scss
@@ -1,155 +1,0 @@
-:root {
-  --content-wrapper-max-width: 1170px;
-}
-
-.container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  * {
-    line-height: 1.6;
-  }
-
-  .wrapper-content {
-    width: 100%;
-    max-width: var(--content-wrapper-max-width);
-    display: flex;
-    padding: 100px 0;
-
-    .img-content {
-      display: flex;
-      padding: 15px;
-      width: 50%;
-      align-items: flex-start;
-
-      .carousel-item {
-        img {
-          height: 100%;
-          width: 100%;
-          border-radius: 8px;
-        }
-      }
-    }
-
-    .text-content {
-      display: flex;
-      flex-direction: column;
-      padding: 15px;
-      width: 50%;
-
-      .product-name {
-        h1 {
-          font-size: 1.7em;
-          font-weight: 600;
-        }
-      }
-
-      .separator {
-        width: 40px;
-        height: 3px;
-        background-color: rgba(128, 128, 128, 0.3);
-        margin: 10px 0;
-      }
-
-      .price {
-        span {
-          color: rgb(192, 31, 65);
-          font-size: 24px;
-        }
-      }
-
-      .sold-quantity {
-        span {
-          text-transform: uppercase;
-          color: rgba(0, 0, 0, 0.7);
-          font-size: 12px;
-        }
-      }
-
-      .description {
-        ul {
-          list-style-type: none;
-          li {
-            margin-bottom: 20px;
-          }
-        }
-      }
-
-      .shopping {
-        display: flex;
-        flex-direction: column;
-        border-top: 1px solid rgba(0, 0, 0, 0.1);
-        padding-top: 15px;
-
-        .tip {
-          color: #7a9c59;
-          span {
-            i {
-              margin-right: 5px;
-              font-size: 1.2em;
-            }
-          }
-        }
-
-        .shopping-action {
-          margin-top: 10px;
-          display: flex;
-          flex-wrap: wrap;
-
-          .quantity-buying {
-            margin-right: 20px;
-            margin-bottom: 20px;
-          }
-
-          .add-to-cart {
-            margin-bottom: 20px;
-            button {
-              padding-top: 5px;
-              padding-bottom: 5px;
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-@media screen and (max-width: 1024px) {
-  .container {
-    .wrapper-content {
-      flex-direction: column;
-      padding: 40px 0;
-
-      .img-content {
-        width: 100%;
-        margin: auto;
-      }
-
-      .text-content {
-        width: 100%;
-
-        .product-name {
-          h1 {
-            font-size: 1.7em;
-          }
-        }
-      }
-    }
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .container {
-    .wrapper-content {
-      .text-content {
-        .product-name {
-          h1 {
-            font-size: 1.5em;
-          }
-        }
-      }
-    }
-  }
-}
