--- src/pages/ProductDetail/components/ProductOverview/components/AddToCartButton/AddToCartButton.module.scss
@@ -1,22 +1,0 @@
-.add-to-cart-btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    background-color: rgba(192, 31, 65, 0.6);
-    border-radius: 99px;
-    border: 1px solid rgba(0, 0, 0, 0.05);
-    padding: 0 15px;
-    transition: all 0.3s;
-    
-    &:hover {
-        background-color: rgba(192, 31, 65, 0.8);
-    }
-
-    span {
-        text-transform: uppercase;
-        font-weight: 700;
-        color: #fff;
-        font-size: 16px;
-    }
-}
