--- src/pages/ProductDetail/components/ProductOverview/components/CountButton/CountButton.module.scss
@@ -1,26 +1,0 @@
-.container {
-  display: flex;
-  align-items: center;
-  border-radius: 99px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-
-  .count-btn {
-    button {
-      padding: 5px 10px;
-      font-size: 16px;
-      transition: all 0.3s;
-
-      &:hover {
-        background-color: rgba(0, 0, 0, 0.1);
-      }
-    }
-
-    input {
-      font-size: 16px;
-      width: 30px;
-      text-align: center;
-      background-color: transparent;
-    }
-  }
-}
