--- src/pages/ProductDetail/components/SimilarProduct/SimilarProduct.js
@@ -1,14 +1,0 @@
-import classNames from "classnames/bind";
-import style from "./SimilarProduct.module.scss"
-
-const cx = classNames.bind(style)
-
-function SimilarProduct() {
-  return (
-    <div>
-      <div></div>
-    </div>
-  );
-}
-
-export default SimilarProduct;
