+++ src/routes/routes.js
@@ -1,19 +1,13 @@
 import Home from "pages/Home/Home";
 import ProductCategory from "pages/ProductCategory/ProductCategory";
-import ProductDetail from "pages/ProductDetail/ProductDetail";
-import routes from "config/routes";
 
 export const publicRoutes = [
   {
-    path: routes.home,
+    path: "/",
     element: <Home />,
   },
   {
     path: "/category/:categoryId",
     element: <ProductCategory />,
-  },
-  {
-    path: "/product/:productId",
-    element: <ProductDetail />,
   },
 ];
